Functions

rel_path_m

MambaModels.rel_path_m — Method.

relpathm

Relative path using the MambaModels src/ directory.

Example to get access to the data subdirectory

rel_path_m("..", "data")

source

generate_m

MambaModels.generate_m — Method.

generate

Utility function to generate all notebooks and chapters from scripts in the scripts directory.

Method

generate_m(sd = script_dict_m)

Required arguments

None, all notebooks/.. and chapters/.. files are regenerated.

source

MambaModels.generate_m — Method.

generate

Generate notebooks and scripts in a single chapter.

Method

generate_m(chapter::AbstractString)

Required arguments

Generate notebooks and scripts in a single chapter, e.g. generate_m("04")

source

MambaModels.generate_m — Method.

generate

Generate a single notebook and script

Method

generate_m(chapter::AbstractString, file::AbstractString)

Required arguments

Generate notebook and script file in chapter, e.g. generatem("04", "m4.1d.jl") or generatem("04/m4.1d.jl")

source

ScriptEntry

StatisticalRethinking.ScriptEntry — Type.

ScriptEntry

Define processing requirements for chapter scripts

Constructor

scriptentry(scriptfile;; nb=true, exe=true, doc=true)

Required arguments

* `scriptfile::AbstractString`        : Script file

Optional arguments

* `nb::Bool`      : Generate a notebook version in notebooks directory
* `exe::Bool`     : Execute the notebook (for testing or documentation purposes)
* `doc::Bool`     : Insert documention into Github pages

If exe = false and doc = true it is assumed the appropriate .md files have been manually created and stored in docs/src/nn/... (Travis will terminate if runs take too long).

scriptentry

StatisticalRethinking.scriptentry — Method.

scriptentry

Constructor for ScriptEntry objects.

Constructor

scriptentry(scriptfile;; nb=true, exe=true, doc=true)

Required arguments

* `scriptfile::AbstractString`        : Script file

Optional arguments

* `nb::Bool`      : Generate a notebook version in notebooks directory
* `exe::Bool`     : Execute the notebook (for testing or documentation purposes)
* `doc::Bool`     : Insert documention into Github pages

If exe = false and doc = true it is assumed the appropriate .md files have been manually created and stored in docs/src/nn/... (Travis will terminate if runs take too long).